+++ tests/test_wrappers.py
@@ -20,7 +20,6 @@ from werkzeug.datastructures import LanguageAccept
 from werkzeug.datastructures import MIMEAccept
 from werkzeug.datastructures import MultiDict
 from werkzeug.datastructures import WWWAuthenticate
-from werkzeug.datastructures.structures import _ImmutableOrderedMultiDict
 from werkzeug.exceptions import BadRequest
 from werkzeug.exceptions import RequestedRangeNotSatisfiable
 from werkzeug.exceptions import SecurityError
@@ -996,24 +995,6 @@ def test_new_response_iterator_behavior():
         assert resp.is_sequence
         resp.stream.write("baz")
         assert resp.response == ["foo", "bar", "baz"]
-
-
-@pytest.mark.filterwarnings("ignore:'OrderedMultiDict':DeprecationWarning")
-def test_form_data_ordering():
-    class MyRequest(wrappers.Request):
-        parameter_storage_class = _ImmutableOrderedMultiDict
-
-    req = MyRequest.from_values("/?foo=1&bar=0&foo=3")
-    assert list(req.args) == ["foo", "bar"]
-    assert list(req.args.items(multi=True)) == [
-        ("foo", "1"),
-        ("bar", "0"),
-        ("foo", "3"),
-    ]
-    assert isinstance(req.args, _ImmutableOrderedMultiDict)
-    assert isinstance(req.values, CombinedMultiDict)
-    assert req.values["foo"] == "1"
-    assert req.values.getlist("foo") == ["1", "3"]
 
 
 def test_values():